Fixed state so that two-arg run() is still a valid entry point
Unfortunately, the changes in this branch so far have just created new
state problems with XProcConfiguration.  The goal was to make a
Main.run() that could safely be reused for multiple pipeline calls, as
compared to the one-shot that Main is today.  Not sure this is taking
me in the right direction....